更新版本库,添加了新的脚本

+++ _shell/OpenSource.sh
@@ -4,13 +4,20 @@
 source "./_shell/init.sh"
 NowPath=${NowPath}
 OutPutPath=${OutPutPath}
+DeployPath="git@github.com:mo7cc/BlogSource.git"
+DeployLocalPath="BlogSource"
 
-## 开源 站点源码 到 git@github.com:mo7cc/BlogSource.git
-SourcePath="git@github.com:mo7cc/BlogSource.git"
+## 判断参数
+desc=$1
+if [ -z "${desc}" ]; then
+  echo -e "\033[31m Err:需要开源说明 \033[0m"
+  exit 1
+fi
+echo "同步: ${desc}"
 
+## 构建源码目录
 rm -rf "${OutPutPath}"
 mkdir "${OutPutPath}"
-
 cp -rf "${NowPath}/src" "${OutPutPath}/src"
 cp -rf "${NowPath}/.gitignore" "${OutPutPath}/"
 cp -rf "${NowPath}/tsconfig.json" "${OutPutPath}/"
@@ -19,17 +26,28 @@ cp -rf "${NowPath}/README.md" "${OutPutPath}/"
 cp -rf "${NowPath}/_shell" "${OutPutPath}/_shell"
 cp -rf "${NowPath}/pnpm-lock.yaml" "${OutPutPath}/"
 
-cd "${OutPutPath}" || exit
-nowTime=$(date +%Y-%m-%d\T%H:%M:%S)
-git init
-git add .
-git commit -m "${nowTime}"
-git remote add origin "${SourcePath}"
-git push -f --set-upstream origin master:main
-echo "同步完成"
-rm -rf "${OutPutPath}/.git"
+# 开始进行发布步骤 OutPutPath -> 远程 DeployPath
+## 拉取远程仓库
+git clone "${DeployPath}"
 
-echo "开源版本发布完成"
-echo "https://github.com/mo7cc/BlogSource"
+## git 本地仓库中 .git 移动到 dist 目录中
+mv "${DeployLocalPath}/.git" "${OutPutPath}/.git"
+
+## 删除本地 git 仓库
+rm -rf "${DeployLocalPath}"
+
+## dist 更名为 本地 git 仓库名
+mv "${OutPutPath}" "${DeployLocalPath}"
+
+# 将仓库推送到远端
+cd "${DeployLocalPath}" || exit
+
+git add . &&
+  git commit -m "${desc}" &&
+  git push &&
+  ## 清理 dist
+  rm -rf ".git" &&
+  mv "${DeployLocalPath}" "${OutPutPath}" &&
+  echo "已推送至${DeployPath}"
 
 exit 0
